flink/requirements.txt
apache-flink==1.17.2
requests
psycopg2-binary
